backend/requirements.txt
asgiref
Django
django-cors-headers
django-rest-framework
djangorestframework-simplejwt
PyJWT
pytz
python-dotenv

pytesseract 
beautifulsoup4
python-docx 
pymupdf
Pillow
google.genai
requests
moviepy 
numpy 
imageio 
imageio-ffmpeg
gradio-client
google.generativeai
gtts
pydub
mutagen